+++ src/room/RTCEngine.ts
@@ -10,6 +10,7 @@ import {
   DataPacket,
   DataPacket_Kind,
   DisconnectReason,
+  ReconnectReason,
   SpeakerInfo,
   TrackInfo,
   UserPacket,
@@ -33,6 +34,7 @@ import {
 import { EngineEvent } from './events';
 import PCTransport, { PCEvents } from './PCTransport';
 import type { ReconnectContext, ReconnectPolicy } from './ReconnectPolicy';
+import CriticalTimers from './timers';
 import type LocalTrack from './track/LocalTrack';
 import type LocalVideoTrack from './track/LocalVideoTrack';
 import type { SimulcastTrackInfo } from './track/LocalVideoTrack';
@@ -118,8 +120,6 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
 
   private clientConfiguration?: ClientConfiguration;
 
-  private connectedServerAddr?: string;
-
   private attemptingReconnect: boolean = false;
 
   private reconnectPolicy: ReconnectPolicy;
@@ -135,6 +135,8 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
   private maxJoinAttempts: number = 1;
 
   private closingLock: Mutex;
+
+  private shouldFailNext: boolean = false;
 
   constructor(private options: InternalRoomOptions) {
     super();
@@ -247,7 +249,13 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
     });
   }
 
-  removeTrack(sender: RTCRtpSender) {
+  /**
+   * Removes sender from PeerConnection, returning true if it was removed successfully
+   * and a negotiation is necessary
+   * @param sender
+   * @returns
+   */
+  removeTrack(sender: RTCRtpSender): boolean {
     if (sender.track && this.pendingTrackResolvers[sender.track.id]) {
       const { reject } = this.pendingTrackResolvers[sender.track.id];
       if (reject) {
@@ -257,9 +265,11 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
     }
     try {
       this.publisher?.pc.removeTrack(sender);
+      return true;
     } catch (e: unknown) {
       log.warn('failed to remove track', { error: e, method: 'removeTrack' });
     }
+    return false;
   }
 
   updateMuteStatus(trackSid: string, muted: boolean) {
@@ -270,8 +280,11 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
     return this.reliableDCSub?.readyState;
   }
 
-  get connectedServerAddress(): string | undefined {
-    return this.connectedServerAddr;
+  async getConnectedServerAddress(): Promise<string | undefined> {
+    if (this.primaryPC === undefined) {
+      return undefined;
+    }
+    return getConnectedAddress(this.primaryPC);
   }
 
   private configure(joinResponse: JoinResponse) {
@@ -313,7 +326,8 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
 
     let primaryPC = this.publisher.pc;
     let secondaryPC = this.subscriber.pc;
-    if (joinResponse.subscriberPrimary) {
+    let subscriberPrimary = joinResponse.subscriberPrimary;
+    if (subscriberPrimary) {
       primaryPC = this.subscriber.pc;
       secondaryPC = this.publisher.pc;
       // in subscriber primary mode, server side opens sub data channels.
@@ -323,11 +337,6 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
     primaryPC.onconnectionstatechange = async () => {
       log.debug(`primary PC state changed ${primaryPC.connectionState}`);
       if (primaryPC.connectionState === 'connected') {
-        try {
-          this.connectedServerAddr = await getConnectedAddress(primaryPC);
-        } catch (e) {
-          log.warn('could not get connected server address', { error: e });
-        }
         const shouldEmit = this.pcState === PCState.New;
         this.pcState = PCState.Connected;
         if (shouldEmit) {
@@ -338,7 +347,12 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
         if (this.pcState === PCState.Connected) {
           this.pcState = PCState.Disconnected;
 
-          this.handleDisconnect('primary peerconnection');
+          this.handleDisconnect(
+            'primary peerconnection',
+            subscriberPrimary
+              ? ReconnectReason.RR_SUBSCRIBER_FAILED
+              : ReconnectReason.RR_PUBLISHER_FAILED,
+          );
         }
       }
     };
@@ -346,7 +360,12 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
       log.debug(`secondary PC state changed ${secondaryPC.connectionState}`);
       // also reconnect if secondary peerconnection fails
       if (secondaryPC.connectionState === 'failed') {
-        this.handleDisconnect('secondary peerconnection');
+        this.handleDisconnect(
+          'secondary peerconnection',
+          subscriberPrimary
+            ? ReconnectReason.RR_PUBLISHER_FAILED
+            : ReconnectReason.RR_SUBSCRIBER_FAILED,
+        );
       }
     };
 
@@ -413,7 +432,7 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
     };
 
     this.client.onClose = () => {
-      this.handleDisconnect('signal');
+      this.handleDisconnect('signal', ReconnectReason.RR_SIGNAL_DISCONNECTED);
     };
 
     this.client.onLeave = (leave?: LeaveRequest) => {
@@ -692,19 +711,19 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
   // websocket reconnect behavior. if websocket is interrupted, and the PeerConnection
   // continues to work, we can reconnect to websocket to continue the session
   // after a number of retries, we'll close and give up permanently
-  private handleDisconnect = (connection: string, signalEvents: boolean = false) => {
+  private handleDisconnect = (connection: string, disconnectReason?: ReconnectReason) => {
     if (this._isClosed) {
       return;
     }
 
-    log.debug(`${connection} disconnected`);
+    log.warn(`${connection} disconnected`);
     if (this.reconnectAttempts === 0) {
       // only reset start time on the first try
       this.reconnectStart = Date.now();
     }
 
     const disconnect = (duration: number) => {
-      log.info(
+      log.warn(
         `could not recover connection after ${this.reconnectAttempts} attempts, ${duration}ms. giving up`,
       );
       this.emit(EngineEvent.Disconnected);
@@ -727,13 +746,14 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
 
     log.debug(`reconnecting in ${delay}ms`);
 
-    if (this.reconnectTimeout) {
-      clearTimeout(this.reconnectTimeout);
-    }
-    this.reconnectTimeout = setTimeout(() => this.attemptReconnect(signalEvents), delay);
+    this.clearReconnectTimeout();
+    this.reconnectTimeout = CriticalTimers.setTimeout(
+      () => this.attemptReconnect(disconnectReason),
+      delay,
+    );
   };
 
-  private async attemptReconnect(signalEvents: boolean = false) {
+  private async attemptReconnect(reason?: ReconnectReason) {
     if (this._isClosed) {
       return;
     }
@@ -753,38 +773,26 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
     try {
       this.attemptingReconnect = true;
       if (this.fullReconnectOnNext) {
-        await this.restartConnection(signalEvents);
+        await this.restartConnection();
       } else {
-        await this.resumeConnection(signalEvents);
+        await this.resumeConnection(reason);
       }
-      this.reconnectAttempts = 0;
+      this.clearPendingReconnect();
       this.fullReconnectOnNext = false;
-      if (this.reconnectTimeout) {
-        clearTimeout(this.reconnectTimeout);
-      }
     } catch (e) {
       this.reconnectAttempts += 1;
-      let reconnectRequired = false;
       let recoverable = true;
-      let requireSignalEvents = false;
       if (e instanceof UnexpectedConnectionState) {
         log.debug('received unrecoverable error', { error: e });
         // unrecoverable
         recoverable = false;
       } else if (!(e instanceof SignalReconnectError)) {
         // cannot resume
-        reconnectRequired = true;
-      }
-
-      // when we flip from resume to reconnect
-      // we need to fire the right reconnecting events
-      if (reconnectRequired && !this.fullReconnectOnNext) {
         this.fullReconnectOnNext = true;
-        requireSignalEvents = true;
       }
 
       if (recoverable) {
-        this.handleDisconnect('reconnect', requireSignalEvents);
+        this.handleDisconnect('reconnect', ReconnectReason.RR_UNKOWN);
       } else {
         log.info(
           `could not recover connection after ${this.reconnectAttempts} attempts, ${
@@ -810,16 +818,14 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
     return null;
   }
 
-  private async restartConnection(emitRestarting: boolean = false) {
+  private async restartConnection() {
     if (!this.url || !this.token) {
       // permanent failure, don't attempt reconnection
       throw new UnexpectedConnectionState('could not reconnect, url or token not saved');
     }
 
     log.info(`reconnecting, attempt: ${this.reconnectAttempts}`);
-    if (emitRestarting || this.reconnectAttempts === 0) {
-      this.emit(EngineEvent.Restarting);
-    }
+    this.emit(EngineEvent.Restarting);
 
     if (this.client.isConnected) {
       await this.client.sendLeave();
@@ -842,6 +848,11 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
       throw new SignalReconnectError();
     }
 
+    if (this.shouldFailNext) {
+      this.shouldFailNext = false;
+      throw new Error('simulated failure');
+    }
+
     await this.waitForPCConnected();
     this.client.setReconnected();
 
@@ -849,7 +860,7 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
     this.emit(EngineEvent.Restarted, joinResponse);
   }
 
-  private async resumeConnection(emitResuming: boolean = false): Promise<void> {
+  private async resumeConnection(reason?: ReconnectReason): Promise<void> {
     if (!this.url || !this.token) {
       // permanent failure, don't attempt reconnection
       throw new UnexpectedConnectionState('could not reconnect, url or token not saved');
@@ -860,12 +871,10 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
     }
 
     log.info(`resuming signal connection, attempt ${this.reconnectAttempts}`);
-    if (emitResuming || this.reconnectAttempts === 0) {
-      this.emit(EngineEvent.Resuming);
-    }
+    this.emit(EngineEvent.Resuming);
 
     try {
-      const res = await this.client.reconnect(this.url, this.token, this.participantSid);
+      const res = await this.client.reconnect(this.url, this.token, this.participantSid, reason);
       if (res) {
         const rtcConfig = this.makeRTCConfiguration(res);
         this.publisher.pc.setConfiguration(rtcConfig);
@@ -879,6 +888,11 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
       throw new SignalReconnectError(message);
     }
     this.emit(EngineEvent.SignalResumed);
+
+    if (this.shouldFailNext) {
+      this.shouldFailNext = false;
+      throw new Error('simulated failure');
+    }
 
     this.subscriber.restartingIce = true;
 
@@ -918,11 +932,6 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
         this.primaryPC?.connectionState === 'connected'
       ) {
         this.pcState = PCState.Connected;
-        try {
-          this.connectedServerAddr = await getConnectedAddress(this.primaryPC);
-        } catch (e) {
-          log.warn('could not get connected server address', { error: e });
-        }
       }
       if (this.pcState === PCState.Connected) {
         return;
@@ -1019,7 +1028,7 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
 
       const negotiationTimeout = setTimeout(() => {
         reject('negotiation timed out');
-        this.handleDisconnect('negotiation');
+        this.handleDisconnect('negotiation', ReconnectReason.RR_SIGNAL_DISCONNECTED);
       }, this.peerConnectionTimeout);
 
       const cleanup = () => {
@@ -1051,7 +1060,7 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
         if (e instanceof NegotiationError) {
           this.fullReconnectOnNext = true;
         }
-        this.handleDisconnect('negotiation');
+        this.handleDisconnect('negotiation', ReconnectReason.RR_UNKOWN);
       });
     });
   }
@@ -1074,20 +1083,28 @@ export default class RTCEngine extends (EventEmitter as new () => TypedEventEmit
     }
   }
 
-  private clearPendingReconnect() {
+  /* @internal */
+  failNext() {
+    // debugging method to fail the next reconnect/resume attempt
+    this.shouldFailNext = true;
+  }
+
+  private clearReconnectTimeout() {
     if (this.reconnectTimeout) {
-      clearTimeout(this.reconnectTimeout);
+      CriticalTimers.clearTimeout(this.reconnectTimeout);
     }
+  }
+
+  private clearPendingReconnect() {
+    this.clearReconnectTimeout();
     this.reconnectAttempts = 0;
   }
 
   private handleBrowserOnLine = () => {
     // in case the engine is currently reconnecting, attempt a reconnect immediately after the browser state has changed to 'onLine'
     if (this.client.isReconnecting) {
-      if (this.reconnectTimeout) {
-        clearTimeout(this.reconnectTimeout);
-      }
-      this.attemptReconnect(true);
+      this.clearReconnectTimeout();
+      this.attemptReconnect(ReconnectReason.RR_SIGNAL_DISCONNECTED);
     }
   };